Update and rename !*!_import_data_and_general_analysis.py to Analysis_GUI_2.0.py
I have redone much of the previous program to improve the syntax and make it more versatile. Changes include: 

Second tk window is now able to allow the user to choose the columns in the db for the Haversine function to utilize. 

Tkinter windows do not have embedded logic loops performing function within them. Widgets also now open and close as intended and do not run forever. 

Added more imports for future expansion.

Added sections for breakdown in IDE, 'Imports' and 'Build'. Currently naming this 'Build 2.0' as much of the internal logic has been modified and changed. 

Added program goal and plan to the top of the Build section.